+++ include/quickSort.h
@@ -1,6 +1,6 @@
 #ifndef QUICKSORT_H
 #define QUICKSORT_H
 
-void quickSort(int *array, int size, int intervalInSeconds);
+void quickSort(int *array, int size, double intervalInSeconds);
 
 #endif // QUICKSORT_H
